requirements.txt
anthropic
openai
google-genai
groq
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv